Removed unnecessary joins
