+++ API/Entities/EDayOff/DayOffController.cs
@@ -3,7 +3,6 @@ using AlpimiAPI.Entities.EDayOff.DTO;
 using AlpimiAPI.Entities.EDayOff.Queries;
 using AlpimiAPI.Locales;
 using AlpimiAPI.Responses;
-using AlpimiAPI.Settings;
 using AlpimiAPI.Utilities;
 using MediatR;
 using Microsoft.AspNetCore.Authorization;
