+++ API/Entities/EGroup/GroupController.cs
@@ -3,7 +3,6 @@ using AlpimiAPI.Entities.EGroup.DTO;
 using AlpimiAPI.Entities.EGroup.Queries;
 using AlpimiAPI.Locales;
 using AlpimiAPI.Responses;
-using AlpimiAPI.Settings;
 using AlpimiAPI.Utilities;
 using MediatR;
 using Microsoft.AspNetCore.Authorization;
