+++ API/Entities/EStudent/StudentController.cs
@@ -3,7 +3,6 @@ using AlpimiAPI.Entities.EStudent.DTO;
 using AlpimiAPI.Entities.EStudent.Queries;
 using AlpimiAPI.Locales;
 using AlpimiAPI.Responses;
-using AlpimiAPI.Settings;
 using AlpimiAPI.Utilities;
 using MediatR;
 using Microsoft.AspNetCore.Authorization;
