+++ API/Entities/ESubgroup/SubgroupController.cs
@@ -3,7 +3,6 @@ using AlpimiAPI.Entities.ESubgroup.DTO;
 using AlpimiAPI.Entities.ESubgroup.Queries;
 using AlpimiAPI.Locales;
 using AlpimiAPI.Responses;
-using AlpimiAPI.Settings;
 using AlpimiAPI.Utilities;
 using MediatR;
 using Microsoft.AspNetCore.Authorization;
