+++ Test/Entities/EClassroom/Commands/UpdateCLassroomCommand.unit.cs
@@ -2,7 +2,6 @@
 using AlpimiAPI.Entities.EClassroom;
 using AlpimiAPI.Entities.EClassroom.Commands;
 using AlpimiAPI.Entities.EClassroomType;
-using AlpimiAPI.Entities.ESchedule;
 using AlpimiAPI.Locales;
 using AlpimiAPI.Responses;
 using AlpimiTest.TestSetup;
