+++ Test/Entities/ELesson/Commands/UpdateLessonCommand.unit.cs
@@ -1,6 +1,4 @@
 ﻿using AlpimiAPI.Database;
-using AlpimiAPI.Entities.EClassroom;
-using AlpimiAPI.Entities.EClassroom.Commands;
 using AlpimiAPI.Entities.EClassroomType;
 using AlpimiAPI.Entities.EGroup;
 using AlpimiAPI.Entities.ELesson;
