+++ Test/Entities/ELessonBlock/Commands/UpdateLessonBlockCommand.unit.cs
@@ -1,21 +1,15 @@
 ﻿using AlpimiAPI.Database;
 using AlpimiAPI.Entities.EClassroom;
-using AlpimiAPI.Entities.EClassroom.Commands;
-using AlpimiAPI.Entities.EClassroomType;
-using AlpimiAPI.Entities.EGroup;
 using AlpimiAPI.Entities.ELesson;
 using AlpimiAPI.Entities.ELessonBlock;
 using AlpimiAPI.Entities.ELessonBlock.Commands;
 using AlpimiAPI.Entities.ELessonType;
-using AlpimiAPI.Entities.ESchedule;
 using AlpimiAPI.Entities.EScheduleSettings;
-using AlpimiAPI.Entities.ESubgroup;
 using AlpimiAPI.Entities.ETeacher;
 using AlpimiAPI.Locales;
 using AlpimiAPI.Responses;
 using AlpimiTest.TestSetup;
 using AlpimiTest.TestUtilities;
-using Microsoft.EntityFrameworkCore.ChangeTracking.Internal;
 using Microsoft.Extensions.Localization;
 using Moq;
 using Newtonsoft.Json;
